GEODE-6425: fix region stat counter rollover
Stats counters used to be integers and have been converted to longs in
order to have fewer issues with rollover. The following stats have been
changed from ints to longs:

* invalidates
* destroys
* txRollbackChanges
* txCommit Changes
* updates
* gets
* misses
* loadsCompleted
* netLoadsCompleted
* netSearchesCompleted
* cacheWriterCallsCompleted
* cacheListenerCallsCompleted
* creates
* puts

+++ geode-core/src/integrationTest/java/org/apache/geode/TXJUnitTest.java
@@ -247,8 +247,8 @@ public class TXJUnitTest {
       cmtre.setUserAttribute("uaValue1");
       assertEquals("uaValue1", cmtre.getUserAttribute());
 
-      int txRollbackChanges = stats.getTxRollbackChanges();
-      int txCommitChanges = stats.getTxCommitChanges();
+      long txRollbackChanges = stats.getTxRollbackChanges();
+      long txCommitChanges = stats.getTxCommitChanges();
       int txFailureChanges = stats.getTxFailureChanges();
       this.txMgr.begin();
       Region.Entry txre = this.region.getEntry("uaKey");
@@ -299,8 +299,8 @@ public class TXJUnitTest {
     }
 
     {
-      int txRollbackChanges = stats.getTxRollbackChanges();
-      int txCommitChanges = stats.getTxCommitChanges();
+      long txRollbackChanges = stats.getTxRollbackChanges();
+      long txCommitChanges = stats.getTxCommitChanges();
       int txFailureChanges = stats.getTxFailureChanges();
       this.region.create("key1", "value1");
       this.txMgr.begin();
@@ -432,7 +432,7 @@ public class TXJUnitTest {
   @Test
   public void testTwoRegionTxs() throws CacheException {
     final CachePerfStats stats = this.cache.getCachePerfStats();
-    int txCommitChanges;
+    long txCommitChanges;
     TransactionId myTxId;
 
     AttributesFactory<String, String> attributesFactory = new AttributesFactory<>();
@@ -2739,7 +2739,7 @@ public class TXJUnitTest {
   }
 
   private void doNonTxInvalidateRegionOp(CachePerfStats stats) throws Exception {
-    int txRollbackChanges = stats.getTxRollbackChanges();
+    long txRollbackChanges = stats.getTxRollbackChanges();
     this.region.create("key1", "value1");
     this.region.create("key2", "value2");
     this.txMgr.begin();
@@ -2771,7 +2771,7 @@ public class TXJUnitTest {
   }
 
   private void doNonTxDestroyRegionOp(CachePerfStats stats) throws Exception {
-    int txRollbackChanges = stats.getTxRollbackChanges();
+    long txRollbackChanges = stats.getTxRollbackChanges();
     this.region.put("key1", "value1");
     this.region.put("key2", "value2");
     this.txMgr.begin();
@@ -4633,9 +4633,9 @@ public class TXJUnitTest {
       private long txCommitTime;
       private long txFailureTime;
       private long txRollbackTime;
-      private int txCommitChanges;
+      private long txCommitChanges;
       private int txFailureChanges;
-      private int txRollbackChanges;
+      private long txRollbackChanges;
 
       private CachePerfStats stats;
 
@@ -4694,7 +4694,7 @@ public class TXJUnitTest {
         this.txRollbackTime = txRollbackTime;
       }
 
-      private void setTxCommitChanges(int txCommitChanges) {
+      private void setTxCommitChanges(long txCommitChanges) {
         this.txCommitChanges = txCommitChanges;
       }
 
@@ -4702,7 +4702,7 @@ public class TXJUnitTest {
         this.txFailureChanges = txFailureChanges;
       }
 
-      private void setTxRollbackChanges(int txRollbackChanges) {
+      private void setTxRollbackChanges(long txRollbackChanges) {
         this.txRollbackChanges = txRollbackChanges;
       }
 
@@ -4831,9 +4831,9 @@ public class TXJUnitTest {
   public void testCacheStats() throws CacheException {
     CachePerfStats cacheStats = this.cache.getCachePerfStats();
     // quick sanity check to make sure perf stats work non-tx
-    int creates;
+    long creates;
     long destroys;
-    int puts;
+    long puts;
     long invalidates;
 
     creates = cacheStats.getCreates();
@@ -5778,7 +5778,7 @@ public class TXJUnitTest {
 
     { // distributed invalidate
       // first make sure invalidate is counted as a change
-      int txRollbackChanges = stats.getTxRollbackChanges();
+      long txRollbackChanges = stats.getTxRollbackChanges();
       this.region.create("key1", "value1");
       this.txMgr.begin();
       this.region.invalidate("key1");
@@ -5837,7 +5837,7 @@ public class TXJUnitTest {
 
     { // local invalidate
       // first make sure invalidate is counted as a change
-      int txRollbackChanges = stats.getTxRollbackChanges();
+      long txRollbackChanges = stats.getTxRollbackChanges();
       this.region.create("key1", "value1");
       this.txMgr.begin();
       this.region.localInvalidate("key1");
