+++ geode-core/src/integrationTest/java/org/apache/geode/cache/ProxyJUnitTest.java
@@ -642,13 +642,13 @@ public class ProxyJUnitTest {
     ExpectedEntryEvent expee = new ExpectedEntryEvent();
     expee.r = r;
     expee.key = "key";
-    int creates = getStats().getCreates();
-    // int puts = getStats().getPuts();
-    // int updates = getStats().getUpdates();
+    long creates = getStats().getCreates();
+    // long puts = getStats().getPuts();
+    // long updates = getStats().getUpdates();
     long destroys = getStats().getDestroys();
     long invalidates = getStats().getInvalidates();
-    int gets = getStats().getGets();
-    int misses = getStats().getMisses();
+    long gets = getStats().getGets();
+    long misses = getStats().getMisses();
 
     r.put("key", "value", cbArg);
     expee.op = Operation.CREATE;
@@ -836,13 +836,13 @@ public class ProxyJUnitTest {
     checkNoCW();
     checkCL(expre);
 
-    int creates = getStats().getCreates();
-    // int puts = getStats().getPuts();
-    // int updates = getStats().getUpdates();
+    long creates = getStats().getCreates();
+    // long puts = getStats().getPuts();
+    // long updates = getStats().getUpdates();
     long destroys = getStats().getDestroys();
     // long invalidates = getStats().getInvalidates();
-    int gets = getStats().getGets();
-    int misses = getStats().getMisses();
+    long gets = getStats().getGets();
+    long misses = getStats().getMisses();
     ExpectedEntryEvent expee = new ExpectedEntryEvent();
     expee.r = r;
     expee.key = "key";
@@ -928,13 +928,13 @@ public class ProxyJUnitTest {
     checkNoTL();
     checkCL(expre);
 
-    int creates = getStats().getCreates();
-    // int puts = getStats().getPuts();
-    // int updates = getStats().getUpdates();
+    long creates = getStats().getCreates();
+    // long puts = getStats().getPuts();
+    // long updates = getStats().getUpdates();
     long destroys = getStats().getDestroys();
     long invalidates = getStats().getInvalidates();
-    // int gets = getStats().getGets();
-    // int misses = getStats().getMisses();
+    // long gets = getStats().getGets();
+    // long misses = getStats().getMisses();
     ExpectedEntryEvent expee = new ExpectedEntryEvent();
     expee.r = r;
     expee.key = "key";
